+++ README.md
@@ -1,4 +1,4 @@
-# V0.3
+# V0.2
 Atlikti darbai:
 # V0.1.1
 1) Sukurta struktūra "Studentas"
@@ -17,6 +17,3 @@ Atlikti darbai:
 # V0.2
 1) Pridėta galimybė nuskaityti duomenis iš failo
 2) Patobulintas skaičių generavimas
-# V0.3
-1) Naudojamos f-jos bei duomenų struktūros sukeltos į atskirus .cpp/.h failus
-2) Kur reikia, panaudotas try-catch išimčių valdymas
