--- V0.2/Balas.h
@@ -1,6 +1,0 @@
-#ifndef BALAS_H
-#define BALAS_H
-
-void balas(std::string tipas, std::vector <studentas>& stud);
-
-#endif
