--- V0.2/Input.h
@@ -1,6 +1,0 @@
-#ifndef INPUT_H
-#define INPUT_H
-
-void input(std::vector <studentas>& stu, std::string ras);
-
-#endif
