--- V0.2/Output.h
@@ -1,6 +1,0 @@
-#ifndef OUTPUT_H
-#define OUTPUT_H
-
-void output(std::vector <studentas> stu);
-
-#endif
