--- V0.2/Studentas.h
@@ -1,15 +1,0 @@
-#ifndef STUDENTAS_H
-#define STUDENTAS_H
-
-//#include <string>
-//#include <vector>
-
-struct studentas
-{
-    std::string vardas;
-    std::string pavarde;
-    std::vector <double> nd;
-    int egz;
-    double galutinis;
-};
-#endif // !STUDENTAS_H
